Refactor templateImage into its own component.
